kafka-services/requirements.txt
python-multipart==0.0.19
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
redis==5.0.1
flask==3.0.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1 
kafka-python 